build: rename Autocode -> HaloDotApi
