+++ app/Jobs/ProcessScrim.php
@@ -5,7 +5,7 @@ namespace App\Jobs;
 use App\Enums\QueueName;
 use App\Models\Game;
 use App\Models\Scrim;
-use App\Services\Autocode\InfiniteInterface;
+use App\Services\HaloDotApi\InfiniteInterface;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
