+++ app/Jobs/PullMatchHistory.php
@@ -6,8 +6,8 @@ namespace App\Jobs;
 
 use App\Enums\QueueName;
 use App\Models\Player;
-use App\Services\Autocode\Enums\Mode;
-use App\Services\Autocode\InfiniteInterface;
+use App\Services\HaloDotApi\Enums\Mode;
+use App\Services\HaloDotApi\InfiniteInterface;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
