+++ app/Models/Game.php
@@ -7,9 +7,9 @@ use App\Enums\Outcome;
 use App\Jobs\PullAppearance;
 use App\Models\Contracts\HasHaloDotApi;
 use App\Models\Pivots\PersonalResult;
-use App\Services\Autocode\Enums\Mode;
-use App\Services\Autocode\Enums\PlayerType;
-use App\Services\Autocode\InfiniteInterface;
+use App\Services\HaloDotApi\Enums\Mode;
+use App\Services\HaloDotApi\Enums\PlayerType;
+use App\Services\HaloDotApi\InfiniteInterface;
 use Carbon\Carbon;
 use Database\Factories\GameFactory;
 use Illuminate\Database\Eloquent\Collection;
