+++ app/Models/Player.php
@@ -12,8 +12,8 @@ use App\Jobs\PullServiceRecord;
 use App\Models\Contracts\HasHaloDotApi;
 use App\Models\Pivots\MatchupPlayer;
 use App\Models\Pivots\PersonalResult;
-use App\Services\Autocode\Enums\Mode;
-use App\Services\Autocode\InfiniteInterface;
+use App\Services\HaloDotApi\Enums\Mode;
+use App\Services\HaloDotApi\InfiniteInterface;
 use App\Services\XboxApi\XboxInterface;
 use App\Support\Image\ImageHelper;
 use App\Support\Session\SeasonSession;
