+++ tests/Mocks/ServiceRecord/MockServiceRecordService.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Tests\Mocks\ServiceRecord;
 
-use App\Services\Autocode\Enums\Filter;
+use App\Services\HaloDotApi\Enums\Filter;
 use Tests\Mocks\BaseMock;
 use Tests\Mocks\Traits\HasErrorFunctions;
